+++ src/format/index.js
@@ -1,8 +1,12 @@
 'use strict'
 
 import num, {numRus} from './num'
+import intWord from './intWord'
+import fileSize from './fileSize'
 
 export {
     num,
     numRus,
+    fileSize,
+    intWord,
 }