+++ src/humanize/index.js
@@ -1,9 +1,0 @@
-'use strict'
-
-import intWord from './intWord'
-import fileSize from './fileSize'
-
-export {
-    fileSize,
-    intWord,
-}